+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/armor/LevelableArmor.java
@@ -2,9 +2,11 @@ package dev.willyelton.crystal_tools.common.levelable.armor;
 
 import dev.willyelton.crystal_tools.CrystalTools;
 import dev.willyelton.crystal_tools.Registration;
+import dev.willyelton.crystal_tools.client.events.RegisterKeyBindingsEvent;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
 import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.common.levelable.LevelableItem;
+import dev.willyelton.crystal_tools.utils.EnchantmentUtils;
 import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -25,6 +27,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.ItemAttributeModifiers;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
 
@@ -88,8 +91,21 @@ public class LevelableArmor extends ArmorItem implements LevelableItem, Equipabl
 
     @Override
     public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> components, TooltipFlag flag) {
-        if (stack.getOrDefault(DataComponents.DISABLE_NIGHT_VISION, false)) {
-            components.add(Component.literal("\u00A7c\u00A7l" + "Night Vision Disabled"));
+        String modeSwitchKey = RegisterKeyBindingsEvent.MODE_SWITCH.getKey().getDisplayName().getString();
+        if (stack.getOrDefault(DataComponents.NIGHT_VISION, false)) {
+            if (stack.getOrDefault(DataComponents.DISABLE_NIGHT_VISION, false)) {
+                components.add(Component.literal("\u00A7c\u00A7l" + "Night Vision Disabled (" + modeSwitchKey + ") To Enable"));
+            } else {
+                components.add(Component.literal("\u00A79" + modeSwitchKey + " To Disable Night Vision"));
+            }
+        }
+
+        if (stack.getOrDefault(DataComponents.FROST_WALKER, false)) {
+            if (EnchantmentUtils.hasEnchantment(stack, Enchantments.FROST_WALKER)) {
+                components.add(Component.literal("\u00A79" + "Shift + " + modeSwitchKey + " To Disable Frost Walker"));
+            } else {
+                components.add(Component.literal("\u00A7c\u00A7l" + "Frost Walker Disabled (Shift + " + modeSwitchKey + ") To Enable"));
+            }
         }
 
         appendLevelableHoverText(stack, components, this, flag);